# 🚀 PerformanceOptimizer Phase 2-3 完全実装