… src/core/usecases/PredictiveAnalyzer.js
… tests/unit/usecases/PredictiveAnalyzer.test.js
… tests/unit/usecases/QualityController.test.js